requirements.txt
selenium==4.16.0
pytest==8.1.1
python-dotenv==1.0.0
webdriver-manager==4.0.1